Export components
By exporting the components, we can take advantage of the powerful tree-shaking feature which will omit the unused parts of code -- making the overall app size even smaller.

Also the components are now named, because the `displayName` component property can be inferred from the `const` expression, but not from the `static` class member. (as seen in the changed snapshots :))